Move migrations to south_migrations and add vanilla Django migrations
